refactor: featuresディレクトリを解体し、appディレクトリに統合
