+++ src/app/admin/_components/part/AdminLayout.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { FC, ReactNode } from "react";
-import { useAuthContext } from "@/features/auth/context/AuthContext";
+import { useAuthContext } from "@/app/_components/functional/AuthContext";
 import { Header } from "./Header";
 
 export const AdminLayout: FC<{ children: ReactNode }> = ({ children }) => {
