+++ src/app/admin/_components/part/Header.tsx
@@ -3,7 +3,7 @@ import { signOut } from "firebase/auth";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { FC } from "react";
-import { useAuthContext } from "@/features/auth/context/AuthContext";
+import { useAuthContext } from "@/app/_components/functional/AuthContext";
 import { auth } from "@/lib/firebase";
 
 export const Header: FC = () => {
