+++ src/app/admin/_components/view/index.tsx
@@ -4,8 +4,8 @@ import Head from "next/head";
 import Link from "next/link";
 import { useState, useEffect } from "react";
 // import { ErrorBoundary } from "react-error-boundary";
+import { AdminLayout } from "@/app/admin/_components/part/AdminLayout";
 import { Book, formatPrice, taxIn } from "@/domain/book";
-import { AdminLayout } from "@/features/admin/components/AdminLayout";
 import { bookRepository } from "@/infrastructure/book";
 
 export default function Admin() {
